Back-end/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gcloud==0.18.3
googleapis-common-protos==1.60.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwcrypto==1.5.0
msgpack==1.0.5
oauth2client==4.1.3
Pillow==10.0.0
protobuf==4.24.1
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3
redis==5.0.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
zope.interface==6.0
